requirements.txt
anyio==3.6.2
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.85.1
google-api-core==2.10.2
google-auth==2.13.0
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-cloud-videointelligence==2.8.3
google-cloud-vision==3.1.4
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.50.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
proto-plus==1.22.1
protobuf==4.21.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
python-multipart==0.0.5
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
